% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liczba_absolwentow.R
\name{liczba_absolwentow}
\alias{liczba_absolwentow}
\title{liczba_absolwentow returns a tibble with number of graduates}
\usage{
liczba_absolwentow(kierunek, stopien = c("1", "2", "JM"),
  liczba_kierunkow = 10)
}
\arguments{
\item{kierunek}{a character value, name of study programe, i.e. 'socjologia'}

\item{stopien}{a vector of character values: '1' - studia pierwszego stopnia,
undergraduate education, '2' - studia drugiego stopnia, complementary
Master's studies, 'JM' - studia jednolite magisterskie, 5-year Master's
studies}

\item{liczba_kierunkow}{a numeric value, use it to limit number of study
programmes to display, default value 10}
}
\value{
a tibble with following columns: study programme, study programme,
  uczelnia, wydzial and number of graduates in 2016
}
\description{
'liczba_absolwentow()' returns a tibble with number of graduates in 2016
grouped by study progammes
}
\examples{
liczba_absolwentow('ekonomia', stopien = c('1', '2'))
}
