% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{ranking}
\alias{ranking}
\title{ranking returns a tibble with study progammes}
\usage{
ranking(kierunek, wskaznik = "wynagrodzenie_brutto", stopien = c("1",
  "2", "JM"), liczba_kierunkow = 10, kolejnosc_malejaca = TRUE)
}
\arguments{
\item{kierunek}{a character value, name of study programe, i.e. 'socjologia'}

\item{wskaznik}{a character value or a vector of character values, allowed
values: "wynagrodzenie_brutto", "wynagrodzenie_wzgledne",
"ryzyko_bezrobocia", "bezrobocie_wzgledne", "czas_poszukiwania_pracy"}

\item{stopien}{a vector of character values: '1' - studia pierwszego stopnia,
undergraduate education, '2' - studia drugiego stopnia, complementary
Master's studies, 'JM' - studia jednolite magisterskie, 5-year Master's
studies}

\item{liczba_kierunkow}{a numeric value, use it to limit number of study
programmes to display, default value 10}

\item{kolejnosc_malejaca}{a boolean, whether to display results sorted
according to selected wskaznik (indicator) in descending order, default
TRUE, sroting relates to selected indicator, wskaznik}
}
\value{
a tibble with following columns: study programme, uczelnia, wydzial,
  selected wskaznik(indicator), study level
}
\description{
'ranking()' returns a tibble with study progammes sorted according to
selected indicator (wskaznik)
}
\examples{
ranking('socjologia')
}
