% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wykres_zarobki_bezrobocie.R
\name{wykres_zarobki_bezrobocie}
\alias{wykres_zarobki_bezrobocie}
\title{wykres_zarobki_bezrobocie returns a scatter plot}
\usage{
wykres_zarobki_bezrobocie(kierunek, stopien = c("1", "2", "JM"))
}
\arguments{
\item{kierunek}{a character value, name of study programe, i.e. 'socjologia'}

\item{stopien}{a vector of character values: '1' - studia pierwszego stopnia,
undergraduate education, '2' - studia drugiego stopnia, complementary
Master's studies, 'JM' - studia jednolite magisterskie, 5-year Master's
studies}
}
\value{
a scatter plot
}
\description{
'wykres_zarobki_bezrobocie()' returns a scatter plot with relative indicator of
income on Y-axis and relative indicator of unemployment on X-axis for
selected study programme and study level, each point represents a study
programme
}
\examples{
wykres_zarobki_bezrobocie('socjologia', stopien = c('1', '2'))
}
