% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wyszukaj_studia_tabela.R
\name{wyszukaj_studia_tabela}
\alias{wyszukaj_studia_tabela}
\title{wyszukaj_studia_tabela returns a tibble with study programmes}
\usage{
wyszukaj_studia_tabela(kierunek, stopien = c("1", "2", "JM"),
  min_wynagrodzenie_brutto = NULL, max_czas_poszukiwania_pracy = NULL)
}
\arguments{
\item{kierunek}{a character value, name of study programe, i.e. 'socjologia'}

\item{stopien}{a vector of character values: '1' - studia pierwszego stopnia,
undergraduate education, '2' - studia drugiego stopnia, complementary
Master's studies, 'JM' - studia jednolite magisterskie, 5-year Master's
studies}

\item{min_wynagrodzenie_brutto}{a numeric value in PLN, avarage incomes in
the first year after graduation}

\item{max_czas_poszukiwania_pracy}{a numeric value in months, avarage time
between graduation and begin of first employment equal or lower than
max_czas_poszukiwania_pracy}
}
\value{
a tibble that contains following columns: study programme, university
  name, faculty name, avarage income of graduates in the first year after
  graduation in PLN, avarage time from graduation to begin of first
  employment in months
}
\description{
wyszukaj_studia_tabela returns a tibble with study programmes that meet
selected criteria: graduates with avarage incomes in the first year after
graduation equal or higher than min_wynagrodzenie_brutto and avarage time
between graduation and begin of first employment equal or lower than
max_czas_poszukiwania_pracy
}
\examples{
wyszukaj_studia_tabela('socjologia', stopien = c('1', '2'),
min_wynagrodzenie_brutto = 3000, max_czas_poszukiwania_pracy = 3)
}
